% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ethiopianToGregorian}
\alias{ethiopianToGregorian}
\title{Convert Ethiopian Date to Gregorian Calendar}
\usage{
ethiopianToGregorian(year = numeric(), month = numeric(), date = numeric())
}
\arguments{
\item{year}{numeric representing Ethiopian year}

\item{month}{numeric representing Ethiopian month}

\item{date}{numeric representing Ethiopian day}
}
\value{
character representation of Gregorian date
}
\description{
This function takes the year, month and day associated with 
a date in the Ethiopian calendar and converts it to a 
date in the Gregorian calendar.
}
