% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{firstDayEthiopian}
\alias{firstDayEthiopian}
\title{First Day of the Ethiopian Year}
\usage{
firstDayEthiopian(year)
}
\arguments{
\item{year}{numeric year to find the first day of}
}
\description{
This function returns the first day of the Ethiopian year
}
