% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{gregorianToEthiopian}
\alias{gregorianToEthiopian}
\title{Convert Gregorian Date to Ethiopian Calendar}
\usage{
gregorianToEthiopian(year = numeric(), month = numeric(), date = numeric())
}
\arguments{
\item{year}{numeric representing Gregorian year}

\item{month}{numeric representing Gregorian month}

\item{date}{numeric representing Gregorian day}
}
\value{
character representation of Ethiopian date
}
\description{
This function takes the year, month and day associated with 
a date in the Gregorian calendar and converts it to a 
date in the Ethiopian calendar.
}
