% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toEthiopian}
\alias{toEthiopian}
\title{Convert Gregorian Date to Ethiopian Calendar}
\usage{
toEthiopian(gregorianDate)
}
\arguments{
\item{gregorianDate}{Date or character representing a date in the Gregorian calendar}
}
\value{
Date from the Ethiopian calendar
}
\description{
This function takes a Gregorian date and converts
it to an Ethiopian date.
}
