% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toGregorian}
\alias{toGregorian}
\title{Convert Ethiopian Date to Gregorian Calendar}
\usage{
toGregorian(ethiopianDate, ...)
}
\arguments{
\item{ethiopianDate}{Date or character representing a date in the Ethiopian calendar}
}
\value{
Date from the Gregorian calendar
}
\description{
This function takes an Ethiopian date and converts
it to a Gregorian date.
}
